refactor(web): remove compact thread mode, leave minimal as the only renderer
Compact mode was the original "card feed" (SpaceTaskCardFeed) — bracket
rails, full tool cards, thinking blocks. Minimal mode (MinimalThreadFeed,
shipped earlier on this branch) replaces it as the production view; the
mode toggle was never wired into the UI and the dual-mode plumbing has
been carrying its weight for nothing.

This commit removes the dead path entirely.

Deleted:
- thread/compact/SpaceTaskCardFeed{.tsx,.test.tsx}
- thread/compact/SpaceSystemInitCard{.tsx,.test.tsx}
- thread/compact/space-task-compact-reducer{.ts,.test.ts}
- thread/SpaceTaskThreadModeToggle.tsx (defined, never imported)
- lib/space-task-thread-config.ts (mode store + localStorage key)

Kept (extracted into a smaller, mode-agnostic module):
- thread/space-task-thread-turns.ts
  Re-homes the two reducer exports MinimalThreadFeed actually uses —
  buildAgentTurns and isUserRow — plus the AgentTurnBlock type
  (renamed from CompactLogicalBlock now that "compact" no longer
  describes anything). Each function gets focused unit tests in
  space-task-thread-turns.test.ts.

Simplified:
- SpaceTaskUnifiedThread.tsx:
  * drops the mode-branching render and renders MinimalThreadFeed
    unconditionally
  * removes the floating "agent-name tag" (it was anchored to compact's
    [data-agent-label] blocks and was redundant in minimal anyway —
    every minimal row carries its own header)
  * drops the unused useMessageMaps call
  * removes findCurrentAgent / shortAgentLabel scroll plumbing
- MinimalThreadFeed.tsx: renames CompactLogicalBlock → AgentTurnBlock and
  imports from the new turns module; refreshes the file-level docstring.
- SpaceTaskUnifiedThread.test.tsx: rewritten around the minimal-only
  render path (no more `mockRenderStyle`, no compact assertions).

Net: 7240 web tests pass, tsc clean, biome happy.
